Add tsc.sh to compiler unit test cases with tsc

+++ src/MapleFE/test/typescript/ts2cxx-test.sh
@@ -39,8 +39,10 @@ for f; do
   ) >& $f-ts2cpp.out &
 done 2>&1 
 wait
-echo -e "\nDate: $(date)\nTest cases passed:" | tee -a $log
-sort ts2cpp.summary.out | xargs -n1 | nl | tee -a $log
+if [ -f ts2cpp.summary.out ]; then
+  echo -e "\nDate: $(date)\nTest cases passed:" | tee -a $log
+  sort ts2cpp.summary.out | xargs -n1 | nl | tee -a $log
+fi
 if [ -f ts2cpp.failures2.out ]; then
   echo -e "\nTest cases failed due to g++ or run:" | tee -a $log
   sort ts2cpp.failures2.out | xargs -n1 | nl | tee -a $log
